refactor: prepare fuzzy extension
